move CurrentUserGuild to http, fix tests
